+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Release date `UNRELEASED`
 
+- (feat api): add `csp/find-directive` function
+
 ## `0.2.1-22`
 
 Release date `2023-03-14`